+++ src/components/carousel/carousel.css
@@ -27,10 +27,8 @@
   border: none;
 }
 
-
 .carousel__container {
-
-  background-color: #fcf8f3
+  background-color: #fcf8f3;
 }
 
 .carousel {
@@ -82,11 +80,13 @@
   border-radius: 25px;
   background: #e89f71;
 }
+
 .line1 {
   position: absolute;
   width: 28px;
   height: 28px;
 }
+
 .line2 {
   position: absolute;
   width: 28px;
@@ -94,6 +94,7 @@
 }
 
 /*CONTAINER PEQUEÑO*/
+
 .pequeño_contenedor {
   display: flex;
   flex-direction: column;
@@ -101,7 +102,6 @@
   position: absolute;
   width: 275px;
   height: 100px;
-  
   left: 630px;
   top: 450px;
   background: rgba(255, 255, 255, 0.6);
@@ -120,6 +120,7 @@
   font-size: 16px;
   color: #616161;
 }
+
 .flecha {
   position: absolute;
   left: 75%;
