+++ src/pages/products/products.css
@@ -1,14 +1,12 @@
-.navbar__container{
+.navbar__container {
     margin-top: 5px;
-  
 }
-
 
 .dropDowns-container {
     display: flex;
     top: 5px;
-   
 }
-.productos__container{
+
+.productos__container {
     margin-top: 30px;
 }
